+++ sdk/data/azappconfig/response_types.go
@@ -6,22 +6,32 @@
 
 package azappconfig
 
-import "time"
+import (
+	"time"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/data/azappconfig/internal/exported"
+)
+
+// SyncToken contains data used in the Sync-Token header.
+// See [Azure App Configuration documentation] for more information on sync tokens.
+//
+// [Azure App Configuration documentation]: https://learn.microsoft.com/azure/azure-app-configuration/rest-api-consistency
+type SyncToken = exported.SyncToken
 
 // AddSettingResponse contains the response from AddSetting method.
 type AddSettingResponse struct {
 	Setting
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }
 
 // DeleteSettingResponse contains the response from DeleteSetting method.
 type DeleteSettingResponse struct {
 	Setting
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }
 
 // GetSettingResponse contains the configuration setting retrieved by GetSetting method.
@@ -31,8 +41,8 @@ type GetSettingResponse struct {
 	// Contains the timestamp of when the configuration setting was last modified.
 	LastModified *time.Time
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }
 
 // ListRevisionsPageResponse contains the configuration settings returned by ListRevisionsPager.
@@ -40,8 +50,8 @@ type ListRevisionsPageResponse struct {
 	// Contains the configuration settings returned that match the setting selector provided.
 	Settings []Setting
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }
 
 // ListSettingsPageResponse contains the configuration settings returned by ListRevisionsPager.
@@ -49,22 +59,22 @@ type ListSettingsPageResponse struct {
 	// Contains the configuration settings returned that match the setting selector provided.
 	Settings []Setting
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }
 
 // SetReadOnlyResponse contains the response from SetReadOnly method.
 type SetReadOnlyResponse struct {
 	Setting
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }
 
 // SetSettingResponse contains the response from SetSetting method.
 type SetSettingResponse struct {
 	Setting
 
-	// Sync token for the Azure App Configuration client, corresponding to the current state of the client.
-	SyncToken *string
+	// SyncToken contains the value returned in the Sync-Token header.
+	SyncToken SyncToken
 }