+++ sdk/data/azappconfig/CHANGELOG.md
@@ -1,15 +1,20 @@
 # Release History
 
-## 0.6.0 (Unreleased)
+## 0.6.0 (2023-09-20)
 
 ### Features Added
 * Handle setting content type in `AddSetting` and `SetSetting` ([#19797](https://github.com/Azure/azure-sdk-for-go/issues/19797))
+* Added type `SyncToken` for better type safety when handling Sync-Token header values.
 
 ### Breaking Changes
 * Response types `ListRevisionsPage` and `ListSettingsPage` now have the suffix `Response` in their names.
+* Method `UpdateSyncToken` on type `Client` has been replaced with `SetSyncToken`.
+* Response types' `SyncToken` field type has changed from `*string` to `SyncToken`.
 
 ### Bugs Fixed
 * Fixed an issue that could cause HTTP requests to fail with `http.StatusUnauthorized` in some cases.
+* The pipeline policy for setting the `Sync-Token` header in HTTP requests now properly formats the value.
+* The caching mechanism for `Sync-Token` values is now goroutine safe.
 
 ### Other Changes
 * `NewClientFromConnectionString()` will return a more descriptive error message when parsing the connection string fails.